feat!: rename block_elements -> elems (#141)
as i steam ahead towards a v1.0.0 release, it's become more and more painful to type block_elements, especially in a crate with such an emphasis on modules being part of the name. This shorter name has been available as an alias for a while, and now the alias is becoming the name of the module.